Added Modal to VIEW
