+++ package.json
@@ -19,6 +19,7 @@
     "msw": "^0.41.0",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
+    "react-draggable": "^4.4.5",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
