+++ src/components/Supervisor/rows.js
@@ -1,0 +1,804 @@
+   //test data
+   export const rows = [
+    {
+
+        id: 1,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+        command: 'test'
+    },
+    {
+
+        id: 2,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Search For Me',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 3,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+
+    , {
+
+        id: 4,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }        , {
+
+        id: 5,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }        , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }   , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }   , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }   , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }   , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }   , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }   , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }   , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }   , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+    , {
+
+        id: 6,
+        date: 'test',
+        location: 'test',
+        incidentType: 'Other',
+        harm: 'test',
+        eventType: 'test',
+        individuals: 'test',
+        effects: 'test',
+        time: 'test',
+        patientSSN: 'test',
+        patientPhone: 'test',
+        patientAddress: 'test',
+        patientName: 'test',
+        witness1Name: 'test',
+        witness1Phone: 'test',
+        witness2Name: 'test',
+        witness2Phone: 'test',
+        witness3Name: 'test',
+        witness3Phone: 'test',
+        department: 'test',
+        description: 'test',
+        prevention: 'test',
+
+        command: 'test'
+    }
+
+];
